# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GroundworkOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: List all jobs
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobRecord"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRecord"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRecord"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRecord"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── QUOTES ───────────────────────────────────────────────────────────────
...
  /documents:
    get:
      operationId: getDocuments
      tags: [documents]
      summary: List all documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentRecord"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRecord"
  /documents/{id}:
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRecord"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── SCHEDULE ─────────────────────────────────────────────────────────────
  /schedule:
    get:
      operationId: getSchedule
      tags: [schedule]
      summary: List schedule entries
      responses:
        "200":
          description: List of schedule entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleEntryRecord"
    post:
      operationId: createScheduleEntry
      tags: [schedule]
      summary: Create a schedule entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleEntryRecord"
  /schedule/{id}:
    patch:
      operationId: updateScheduleEntry
      tags: [schedule]
      summary: Update a schedule entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleEntryRecord"
    delete:
      operationId: deleteScheduleEntry
      tags: [schedule]
      summary: Delete a schedule entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── PLANT ────────────────────────────────────────────────────────────────
  /plant:
    get:
      operationId: getPlant
      tags: [plant]
      summary: List all plant items
      responses:
        "200":
          description: List of plant items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlantRecord"
    post:
      operationId: createPlantItem
      tags: [plant]
      summary: Create a plant item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantRecord"
  /plant/{id}:
    patch:
      operationId: updatePlantItem
      tags: [plant]
      summary: Update a plant item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlantUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlantRecord"
    delete:
      operationId: deletePlantItem
      tags: [plant]
      summary: Delete a plant item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── RATE BOOK ────────────────────────────────────────────────────────────
  /rate-book:
    get:
      operationId: getRateBook
      tags: [rate-book]
      summary: List rate book entries
      responses:
        "200":
          description: List of rate book entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RateBookRecord"
    post:
      operationId: createRateBookEntry
      tags: [rate-book]
      summary: Create a rate book entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateBookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateBookRecord"
  /rate-book/{id}:
    patch:
      operationId: updateRateBookEntry
      tags: [rate-book]
      summary: Update a rate book entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateBookUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateBookRecord"
    delete:
      operationId: deleteRateBookEntry
      tags: [rate-book]
      summary: Delete a rate book entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    JobRecord:
      type: object
      required:
        [
          id,
          jobNumber,
          title,
          status,
          progressPercent,
          nrswaRequired,
          createdAt,
        ]
      properties:
        id: { type: string }
        jobNumber: { type: string }
        title: { type: string }
        clientId: { type: ["string", "null"] }
        clientName: { type: ["string", "null"] }
        type: { type: ["string", "null"] }
        siteAddress: { type: ["string", "null"] }
        value: { type: ["number", "null"] }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        status: { type: string }
        progressPercent: { type: integer }
        description: { type: ["string", "null"] }
        foreman: { type: ["string", "null"] }
        crewCount: { type: ["integer", "null"] }
        nrswaRequired: { type: boolean }
        permitNumber: { type: ["string", "null"] }
        createdAt: { type: string }
    JobInput:
      type: object
      required: [id, jobNumber, title]
      properties:
        id: { type: string }
        jobNumber: { type: string }
        title: { type: string }
        clientId: { type: string }
        type: { type: string }
        siteAddress: { type: string }
        value: { type: number }
        startDate: { type: string }
        endDate: { type: string }
        status: { type: string }
        progressPercent: { type: integer }
        description: { type: string }
        foreman: { type: string }
        crewCount: { type: integer }
        nrswaRequired: { type: boolean }
        permitNumber: { type: string }
    JobUpdate:
      type: object
      properties:
        title: { type: string }
        clientId: { type: string }
        type: { type: string }
        siteAddress: { type: string }
        value: { type: number }
        startDate: { type: string }
        endDate: { type: string }
        status: { type: string }
        progressPercent: { type: integer }
        description: { type: string }
        foreman: { type: string }
        crewCount: { type: integer }
        nrswaRequired: { type: boolean }
        permitNumber: { type: string }

    # ── QUOTE ───────────────────────────────────────────────────────────────
...
    DocumentRecord:
      type: object
      required: [id, name, type, status, relatedTo, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        status: { type: string }
        expiryDate: { type: ["string", "null"] }
        issuedDate: { type: ["string", "null"] }
        relatedTo: { type: string }
        relatedId: { type: ["string", "null"] }
        relatedName: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    DocumentInput:
      type: object
      required: [id, name, type, relatedTo]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        status: { type: string }
        expiryDate: { type: string }
        issuedDate: { type: string }
        relatedTo: { type: string }
        relatedId: { type: string }
        relatedName: { type: string }
        notes: { type: string }
    DocumentUpdate:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        status: { type: string }
        expiryDate: { type: string }
        issuedDate: { type: string }
        relatedTo: { type: string }
        relatedId: { type: string }
        relatedName: { type: string }
        notes: { type: string }

    # ── SCHEDULE ────────────────────────────────────────────────────────────
    ScheduleEntryRecord:
      type: object
      required:
        [id, title, startDatetime, endDatetime, crewCount, type, createdAt]
      properties:
        id: { type: string }
        jobId: { type: ["string", "null"] }
        jobNumber: { type: ["string", "null"] }
        jobTitle: { type: ["string", "null"] }
        clientName: { type: ["string", "null"] }
        title: { type: string }
        startDatetime: { type: string }
        endDatetime: { type: string }
        crewCount: { type: integer }
        plantAssigned: { type: ["string", "null"] }
        foreman: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        type: { type: string }
        createdAt: { type: string }
    ScheduleEntryInput:
      type: object
      required: [id, title, startDatetime, endDatetime, crewCount, type]
      properties:
        id: { type: string }
        jobId: { type: string }
        title: { type: string }
        startDatetime: { type: string }
        endDatetime: { type: string }
        crewCount: { type: integer }
        plantAssigned: { type: string }
        foreman: { type: string }
        notes: { type: string }
        type: { type: string }
    ScheduleEntryUpdate:
      type: object
      properties:
        jobId: { type: string }
        title: { type: string }
        startDatetime: { type: string }
        endDatetime: { type: string }
        crewCount: { type: integer }
        plantAssigned: { type: string }
        foreman: { type: string }
        notes: { type: string }
        type: { type: string }

    # ── PLANT ───────────────────────────────────────────────────────────────
    PlantRecord:
      type: object
      required: [id, name, category, status, owned, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        registration: { type: ["string", "null"] }
        category: { type: string }
        make: { type: ["string", "null"] }
        model: { type: ["string", "null"] }
        year: { type: ["integer", "null"] }
        status: { type: string }
        currentJobId: { type: ["string", "null"] }
        currentJobTitle: { type: ["string", "null"] }
        serviceDue: { type: ["string", "null"] }
        motDue: { type: ["string", "null"] }
        thoroughExamDue: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        dailyRate: { type: ["number", "null"] }
        owned: { type: boolean }
        createdAt: { type: string }
    PlantInput:
      type: object
      required: [id, name, category]
      properties:
        id: { type: string }
        name: { type: string }
        registration: { type: string }
        category: { type: string }
        make: { type: string }
        model: { type: string }
        year: { type: integer }
        status: { type: string }
        currentJobId: { type: string }
        serviceDue: { type: string }
        motDue: { type: string }
        thoroughExamDue: { type: string }
        notes: { type: string }
        dailyRate: { type: number }
        owned: { type: boolean }
    PlantUpdate:
      type: object
      properties:
        name: { type: string }
        registration: { type: string }
        category: { type: string }
        make: { type: string }
        model: { type: string }
        year: { type: integer }
        status: { type: string }
        currentJobId: { type: string }
        serviceDue: { type: string }
        motDue: { type: string }
        thoroughExamDue: { type: string }
        notes: { type: string }
        dailyRate: { type: number }
        owned: { type: boolean }

    # ── RATE BOOK ───────────────────────────────────────────────────────────
    RateBookRecord:
      type: object
      required:
        [
          id,
          category,
          description,
          unit,
          labourRate,
          materialRate,
          plantRate,
          totalRate,
          createdAt,
        ]
      properties:
        id: { type: string }
        category: { type: string }
        description: { type: string }
        unit: { type: string }
        labourRate: { type: number }
        materialRate: { type: number }
        plantRate: { type: number }
        totalRate: { type: number }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    RateBookInput:
      type: object
      required: [id, category, description, unit]
      properties:
        id: { type: string }
        category: { type: string }
        description: { type: string }
        unit: { type: string }
        labourRate: { type: number }
        materialRate: { type: number }
        plantRate: { type: number }
        totalRate: { type: number }
        notes: { type: string }
    RateBookUpdate:
      type: object
      properties:
        category: { type: string }
        description: { type: string }
        unit: { type: string }
        labourRate: { type: number }
        materialRate: { type: number }
        plantRate: { type: number }
        totalRate: { type: number }
        notes: { type: string }